official solution to t9 using mapped operations
You can use a map to specify operations on data sets
for batch operations, the json type can be used to
specify the operation (put, get, del) and the following
fields are the actual data to manipulate.

The key can be anything and the value can apparently
be parsed from the entire json row...  I'm not sure
how it does this or why that is the functionality.